+++ .github/workflows/autotest.yml
@@ -1,7 +1,6 @@
 name: Build and Test
 
 on:
-  push:
   pull_request:
 
 jobs:
