--- src/main/java/com/Ticketease/TE/main/MainController.java
@@ -1,16 +1,0 @@
-package com.Ticketease.TE.main;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@RequiredArgsConstructor
-@Controller
-public class MainController {
-
-    @GetMapping("/")
-    public String index() {
-        return "index";
-    }
-
-}
