--- src/main/java/com/Ticketease/TE/member/Login/LoginController.java
@@ -1,13 +1,0 @@
-package com.Ticketease.TE.member.Login;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class LoginController {
-
-    @GetMapping("/login")
-    public String Login(){
-        return "login";
-    }
-}
