--- src/main/java/com/Ticketease/TE/member/MemberRepository.java
@@ -1,11 +1,0 @@
-package com.Ticketease.TE.member;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-    Optional<Member> findByNickName(String nickname);
-
-}
