+++ src/main/java/com/ticketease/te/DemoApplication.java
@@ -1,0 +1,13 @@
+package com.ticketease.te;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DemoApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DemoApplication.class, args);
+	}
+
+}
