+++ src/main/java/com/ticketease/te/main/MainController.java
@@ -1,0 +1,16 @@
+package com.ticketease.te.main;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@RequiredArgsConstructor
+@Controller
+public class MainController {
+
+    @GetMapping("/")
+    public String index() {
+        return "index";
+    }
+
+}
