+++ src/main/java/com/ticketease/te/member/Login/LoginController.java
@@ -1,0 +1,13 @@
+package com.ticketease.te.member.Login;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class LoginController {
+
+    @GetMapping("/login")
+    public String Login(){
+        return "login";
+    }
+}
