+++ src/main/java/com/ticketease/te/member/Login/LoginRequest.java
@@ -1,0 +1,11 @@
+package com.ticketease.te.member.Login;
+
+public record LoginRequest(
+        String nickname,
+        String password
+) {
+
+    public static LoginRequest of(String nickname, String password){
+        return new LoginRequest(nickname,password);
+    }
+}
