+++ src/main/java/com/ticketease/te/member/MemberRepository.java
@@ -1,0 +1,11 @@
+package com.ticketease.te.member;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+    Optional<Member> findByNickName(String nickname);
+
+}
