+++ src/main/java/com/ticketease/te/member/signUp/SignUpController.java
@@ -1,0 +1,45 @@
+package com.ticketease.te.member.signUp;
+
+import com.ticketease.te.exception.ExceptionCode;
+import com.ticketease.te.exception.SignUpExceptionHandler;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
+@RequiredArgsConstructor
+@Controller
+public class SignUpController {
+
+    private final SignUpService signUpService;
+
+    @GetMapping("/signup")
+    public String SignUpDefaultPage(){
+        return "signup";
+    }
+
+    @PostMapping("/signup")
+    public String signUpUser(@Valid SignUpRequest request,
+                             BindingResult bindingResult,
+                             RedirectAttributes redirectAttributes){
+        try{
+            signUpUserValidation(SignUpRequest.toDto(request),bindingResult);
+            signUpService.signUpUser(SignUpRequest.toDto(request));
+        }catch (SignUpExceptionHandler handler){
+            redirectAttributes.addFlashAttribute(handler.getCode().toString(), handler.getMessage());
+            return "redirect:signup";
+        }
+
+        redirectAttributes.addFlashAttribute("success", "회원가입이 완료됐습니다.");
+        return "redirect:login";
+    }
+
+    private void signUpUserValidation(SignUpDto dto, BindingResult result){
+        if(result.hasErrors() || !dto.password().equals(dto.confirmPassword())){
+            throw new SignUpExceptionHandler(ExceptionCode.INVALID_SIGNUP_REQUEST,ExceptionCode.INVALID_SIGNUP_REQUEST.getDescription());
+        }
+    }
+}
