--- src/test/java/com/Ticketease/TE/DemoApplicationTests.java
@@ -1,13 +1,0 @@
-package com.Ticketease.TE;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DemoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
