+++ src/test/java/com/ticketease/te/DemoApplicationTests.java
@@ -1,0 +1,13 @@
+package com.ticketease.te;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DemoApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
